refactor(core): move QuietHours into its own file
